requirements.txt
# Core Dependencies
numpy>=1.19.0
pandas>=1.2.0
scikit-learn>=0.24.0
scipy>=1.6.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Machine Learning Models
xgboost>=1.3.0

# Data Processing
openpyxl>=3.0.0
pyyaml>=5.4

# Utilities
tqdm>=4.60.0
joblib>=1.0.0
psutil>=5.8.0

# Optional: Deep Learning (RealMLP)
# Uncomment if you want to use RealMLP models
# torch>=1.10.0
# pytorch-lightning>=1.5.0

# Development (Optional)
# pytest>=6.2.0
# black>=21.5b0
# flake8>=3.9.0
